Update index.html
04:34 PM, 09/24/2023 - Alexander Garofalo

Changed out the original content of the main table to contain a placeholder version of what it might look like with a student's information.

Added additional features such as a heading bar, similar to what is in Degree Audit. While this heading bar may be removed later, it is still a work in progress.

Commented out larger portions of the original .html file. This is temporary. Commented content will either be updated with relevant info. or will be removed entirely.