Python Script to filter JSON file
11:59 AM, 10/12/2023 - Alexander Garofalo

There is a lot more content that I need to commit, but I will be doing that in a different commit. This current one pertains to a Python script that I have been working on the past 2 days. The goal is to be able to go to a URL where the data is in a JSON-like format, save the data to a file, then filter out only the desired data and storing this to a new file.

Some of the commented out code pertains to Cookies and grabbing cookie IDs. I was trying to grab my own cookie so the program can run without issues for the user, but it seemed that the cookies the script was grabbing were not "real" in the sense that they were not authenticated with the web server. The acquired cookies had the same format, but they just would not work.