Created /frontend folder
03:26 PM, 10/13/2023 - Alexander Garofalo

I used the command "npx create-react-app frontend" to create the /frontend folder for the React Frontend. I am working on creating a React frontend of the Smart Degree Audit project while keeping the backend as Python Flask.

I have reworked the structure of the directory tree for the React frontend portion, as I separate the .js files into a "components" folder and the .css files into a "styles" folder.

At first, I had the entire index.html file from before (the one being used with the Flask web app) getting returned in App.js, but I found that the rendering process was taking too long. So, this commit shows the work-in-progress as I try to separate out the content into multiple .js component files in the hopes of improving performance.