Update /frontend folder (React)
06:44 PM, 10/14/2023 - Alexander Garofalo

I have made modifications to the styling of the React frontend. It is together, for the most part.

My current goal is to create HTML tables that are always visible, but there is a button that when clicked, will expand the entire table to show everything within it. Then, when the button is clicked again, it reverts to its normal size.

However, I may end up making the tables a static size that cannot be resized at all. It is a decision between complexity, but granting the user more control over their environment, and simplicity, with a rigid environment.