Create various files for new Catalog
10:43 AM, 11/25/2023 - Alexander Garofalo

sdaCatalog.json is going to be the catalog that I intend to use for the project. It will be based on newCatalog.json, but will have various elements removed as I do not see them as having purpose.

newCatalog.json is a copy of 2024SpringCatalog.json.

reCat.py is the file that I used to create sdaCatalog.json.