Update app.py
07:45 PM, 12/03/2023 - Alexander Garofalo

I have managed to connect the MySQL database to the Academic Progress Tracker. I am contemplating removing the "/add_course" route since a student should not be adding courses to it. It should be added automatically on the back-end.

I made it so that the password to access the databases is retrieved from a file, so that the password is not explicitly stored in the code. MySQL Injections are scary.

I also made it so that the MySQL queries are running "prepared statements" to prevent SQL Injections.