Create faculty.json
09:07 AM, 12/04/2023 - Alexander Garofalo

Going to create a list of all the faculty at Stetson University and then create a table in the database for it. I was thinking about this last night: We could use it with the Course Review feature.

Basically, the table gets loaded into a dropdown box that contains all the faculty, so the student, when submitting a course review, can select the professor they had for the course.