Create review.py and reviewsDB.py
09:17 AM, 12/04/2023 - Alexander Garofalo

I am working on creating a script that will auto-create the MySQL table to store the reviews in. Basically, it will automatically format it with the columns, so it is less manual work on my end.

This is basically what I was doing with the sdaCatalog and the other MySQL databases and tables.