Update app.py (Course Review)
09:35 AM, 12/04/2023 - Alexander Garofalo

I added the MySQL connectors to the file to allow it to connect to the database, just like how I did with the Academic Progress Tracker.

I also adjusted the spacing for the form fields to make all the equal signs line up with the rest of its respective cluster. I figured that this would help increase the code's readability. In case you all don't like it, I copied the code and placed the original arrangements as comments below.