Updated path to ourPySQL.txt
09:44 AM, 12/04/2023 - Alexander Garofalo

I moved the ourPySQL.txt file into a folder that is set in the .gitignore file to be excluded from being pushed. I didn't want to keep running the risk of accidently pushing the file to the GitHub repository since it contains the password for the MySQL Databases. Plus, the databases are located locally on my machine.

So, I had to adjust the path for ourPySQL.txt in the app.py file for the Academic Progress Tracker, the Course Review, and the Anonymous. I kept the original line of code for the original path as a comment for ease of use in the future.