Update CourseTemplate.json
10:06 AM, 12/04/2023 - Alexander Garofalo

I had originally created this file to contain the templates for the various setups that we would be using for the MySQL Databases. I am still working on this.

I have completed the template for Courses and how they would be loaded into the sdaCatalog.

I created a template for the Academic Progress Tracker that defines the information that is stored into the MySQL Database for this feature.

I created a template for the Course Review feature and how the reviews would be saved into the MySQL Database.