Create APT folder in main templates folder
10:01 PM, 12/05/2023 - Alexander Garofalo

I created an APT folder inside the main /templates folder for weblet.py. I have moved the progress_tracker.html file into this APT folder.

My plan is to separate the html files for each feature into a folder that is named as an abbreviation of that feature. (i.e. APT stands for "Academic Progress Tracker")