Update style.css
07:25 PM, 12/08/2023 - Alexander Garofalo

I added to the color schemes that Sydney had created. I also resolved the problem that was causing the HTML content to overlap with the Footer.
Problem: The <main> tag was set to "height: 50vh;".

Removed a lot of obsolete styling. Reorganized the styling into new sections. Created sections of styling, breaking them down into what they are going to be focused on.

Styling for the Header, Footer, and Academic Progress Tracker are "DONE". Will run through them again at the end if there is time.