Blueprints: (8) Fixed routes
03:13 PM, 12/10/2023 - Alexander Garofalo

Commit #8 of fixing blueprints.

I have adjusted the URL redirects and page links so that every link works. I eliminated the redundancy that was present with the original setup.

I also reformatted the code for how the parameters to the Blueprints were being declared. I tried to make it more readable, clear, and consistent.