Created special 405 page and APT edits
04:51 PM, 12/10/2023 - Alexander Garofalo

I created a modified "405 Method Not Allowed" page for when the Academic Progress Tracker is accessed with a "GET" request.

I moved the Student Heading that was originally on the home page to the Academic Progress Tracker page. The user must submit data through the ID Form on the home page, which will redirect them to the Progress Tracker, where they can see data about themselves in the Student Heading and their Course History in the table.